requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pydantic==2.0.2
pydantic_core==2.1.2
python-dotenv==1.0.1
requests==2.31.0
typing_extensions==4.12.2
urllib3==2.3.0
